oeqa/selftest/context: Enable support for threaded runs
Add an option to specify how many threads will be used for
execution, default to 1.

Add OE_SELFTEST_THREAD_NUM environment variable for be able
to set thread num into the Yocto Autobuilder and don't need
to figure out the version of oe-selftest script.

If the thread_num are greater than 1 the OESelftestContextThreaded
will be used, this is due to compatibility reasons.

Signed-off-by: Aníbal Limón <anibal.limon@linux.intel.com>