documentation: Moved "Yocto Project Source Repositories" section
Fixes [YOCTO #11630]

The "Yocto Project Source Repositories" section is purely conceptual
and thus does not belong in the reconstituted dev-manual, which is
converting to a "how-to" manual.  I have moved this section to the
ref-manual under "The Yocto Project Development Environment" section.

Moving the section affected several things.  Three figures that where
in the dev-manual were copied and moved to the figures folder of the
ref-manual.  This required deleting them from the dev-manual and
adding them to the ref-manual.

The makefile had to be edited to move those figures into the ref-manual
and delete them from the latest version of the dev-manual in order for
the tarball of files to be correct when the manuals are made.

Links to anchors within the moved section had to be fixed in various
manuals.  Links to the main section "id" had to be fixed to point
to the ref-manual manual.

(From yocto-docs rev: c619ae2bbc6be93cb0324ca44e936eefec12b28c)

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>