yocto-project-qs: Updated "Welcome" section.
I moved the bit about using a VM if you don't have a Linux
machine to a note as it broke the natural flow of this section.
Also, inserted a link to the comprehensive list of links and
other documentation references to the existing list of
other information.

(From yocto-docs rev: 73750063580ef7bc20bd445adc3e2b0b01e2257f)

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>