oe-init-build-env-memres: swap parameter order
This script expected the port number to be specified first, which is
somewhat counter-intuitive especially if you're used to
oe-init-build-env; besides, in local usage you are unlikely to need to
specify a custom port. Given that few people are using this yet (based
on the issues I have recently fixed), switch the arguments around so
that the two scripts behave consistently.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>