initramfs-framework: include install module in recipe for installation
Utilized the existing init-install-efi.sh script to manage the
installation process of images in initramfs-framework model. This
script will be executed when "install" option is being chosen in
the grub menu and install the image on the target platform. A
new install module is being added in the recipe to handle the
installation process using initramfs-framework.

[YOCTO #10989]

Signed-off-by: Ng, Wei Tee <wei.tee.ng@intel.com>