[meta-openwrt] Juci: fix missing Juci website
When building Juci, the www directory was empty. This commit helps to
fix the missing configurations needed by Makefile due to this being a
build from OpenWRT (which depends on menuconfig, Kconfig) and resolve a
bug with the Makefile that does not allow "make -j8" so forced to build
using "make -j1".

Signed-off-by: Stanley Cheong Kwan, Phoong <stanley.cheong.kwan.phoong@intel.com>