bb/main.py: read port and host from bitbake.lock
Fixed when connect to server, for example:
$ bitbake --status-only
or:
$ bitbake -m

It needs read host and port from bitbake.lock in these cases, otherwise it
doesn't know where to connect.

And also change bitbake.lock's content to: "host: port" rather than
"pid host: port", we don't use the pid by now, so not save it can make it easier
when read it.

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>