meta-yocto-bsp: Remove superfluous kernel set
The genericx86 and genericx86-64 machine confs do not need to set the
PREFERRED_PROVIDER_virtual/kernel as they will use the one set by the
distribution. If there needed to be a different version for one of these
machines in the future, we could add it in to the machine.conf.

Signed-off-by: brian avery <brian.avery@intel.com>