meta-yocto-bsp: Remove superfluous kernel set
The machine confs do not need to set the
PREFERRED_PROVIDER_virtual/kernel or PREFERRED_VERSION as they will use
the one set by the distribution. If there needs to be a different
kernel or version for one of these machines in the future, we should add
it to the machine.conf.

(From meta-yocto rev: 89dd8159fc83f54d7fe234b619a28f5d97519156)

Signed-off-by: brian avery <brian.avery@intel.com>