meta: remove kernel settings from machine confs
We set PREFERRED_VERSION and PREFERRED_PROVIDER in various machine conf
files. This necessitates changing them all in sync when we change
linux-yocto kernel versions.  This patch removes those settings so that
the machines will use the ones set in the distribution conf.  Should a
particular machine need to override the kernel or version, that setting
should go into that machines configuration file.

Signed-off-by: brian avery <brian.avery@intel.com>