kernel: make kernel configuration fragments available without kernel-yocto
Although kernel-yocto has always supported kernel configuration
fragments it is not a default capability available to existing recipes
that inherit core kernel support.

The preparation and audit phases of kernel-yocto add processing time and
structure to a kernel that is not always wanted, as such kernel
configuration support has remained outside of the core kernel classes.

kernel-yocto processing has been simplified, and as such, it is possible
to move the mainline merging of kernel configuration fragment into the
core kernel class, while leaving the preparation and audit of kernel
configuration fragments in kernel-yocto. This makes the core support
available to any kernel recipe.

To make this new capability as lightweight as possible, a new task
do_merge_config() is created, and is only enabled if configuration
fragments are detected in the SRC_URI.

do_merge_config has the following properties/variables:

 KCONFIG_MODE: can be 'allnoconfig' or 'alldefconfig' (default)
               The kconfig mode is used to instruct merge_config.sh
               on how unspecified options should be populated.

 KERNEL_CONFIG_FRAGMENTS: lists the fragments that should be
                          merged. This is filled in automatically
                          for anything specified on the SRC_URI.

 ${WORKDIR}/kernel-cfgs: this file lists additional fragments that
                         should be merged. This exists so that wrapping
                         tasks (like linux-yocto) can specify fragments
                         without needing to use anonymous python and
                         variable overrides (which are not always
                         possible).

Also note that once merge_config support is enabled, the existing
KERNEL_CONFIG_COMMAND is overridden, since merge_config takes care of
creating the final .config

linux-yocto is adapted to use this new routine, and results in the
same final .configs as were previously generated.

[YOCTO #8191]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>