scriptutils: fix to use lowercase dummy recipe name
recipetool create (and hence devtool add) and devtool upgrade use
fetch_url() which creates a dummy recipe in order to fetch source.
Previously the random part of the name was using uppercase characters,
and this is shortly going to trigger a QA warning, so use lowercase
instead as I really should have in the first place.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>