image_types: Remove ELF Image Type
the mkelfimage utiltiy was removed from the upstream coreboot
project in 2014. The commit message then mentions it had not
been needed for year. A possible replacement could be cbfstool
which if needed could be enabled in 2.5

This removes the QA test as well as the image_types code.

Additionally the tools to generate the convert.bin.c are
designed for an i386 based compiler, not our current 64 bit
compiler, and would need additional modifications to work
correctly.  We can generate the correct on an i386 target and
then copy it back to the host.

Makes the [YOCTO #11967] bug go away.

Signed-off-by: Saul Wold <sgw@linux.intel.com>