devtool: extract: check the repository and error out if not present
If a bug is present it's possible that the source tree is not a git
repository and thus if we tried to perform git operations (such as
committing or changing branches) when extracting source, then we might
in fact be running those operations on the metadata repository if the
build directory is underneath, say, poky or OE-Core, and that could
make a mess. Check if the source tree is a git repository and refuse
to continue if it isn't.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>