bzip2: Create Makefile for run-ptest
To improve binary reproducibility, only the
relevent part of the Makefile in bzip2-ptest
package that run-ptest needs are copied to
the destination directory.

bzip2-ptest requires Makefile. The original
Makefile contains references to build host.
By copying only the relevent part,
we avoid those references making their way
into the target system

Used Makefile.am as the source to extract
the revelent part instead of Makefile as
it is easier to parse than Makefile.

A redundant command  in do_install_ptest()
that repeat the copying of Makefile that
was already copied by  Makefile's install-ptest
step is removed because it now interfers with
the creation of the new Makefile by
make's install-ptest.

[YOCTO #11596]

Signed-off-by: Ooi Cinly <cinly.ooi@intel.com>