toaster: catch early git and server exceptions
Stop the pending build and report to the user if there is
an internal exception due to a git error, a bitbake server
error, or if the server-only mode halts due to for example
a user syntax error in a layer or recipe.

These exceptions were not caught because they occured before
the normal toastergui processing was started.

[YOCTO #12056]

Signed-off-by: David Reyna <David.Reyna@windriver.com>