signing: export GPG_TTY prior to gpg launch
According to the gpg-agent man page, the GPG_TTY variable should be set to the output
of 'tty' (which is also added as part of HOSTTOOLS). Without this variable,
on gpg versions which >=2.1, rpm signing fails leading to hangs. The symptoms of this
hang is that oe-selftest never finishes (because do_package_write_rpm never finishes)
leaving several gpg zombie process in the system.

[YOCTO #12022]

Signed-off-by: Leonardo Sandoval <leonardo.sandoval.gonzalez@linux.intel.com>